add base epoch manager, prepare for decentralized epoch manager
