+++ src/include/wire/libevent_thread.h
@@ -85,12 +85,13 @@ class LibeventWorkerThread : public LibeventThread {
   LibeventWorkerThread(const int thread_id);
 };
 
+// a master thread contains multiple worker threads.
 class LibeventMasterThread : public LibeventThread {
  private:
   const int num_threads_;
 
   // TODO: have a smarter dispatch scheduler
-  int next_thread_id_ = 0;  // next thread we dispatched to
+  std::atomic<int> next_thread_id_;  // next thread we dispatched to
 
  public:
   LibeventMasterThread(const int num_threads, struct event_base *libevent_base);
